Add chunking support for SPI reads and writes
Linux limits the maximum SPI read and write length based on an internal buffer.
If any individual read or write exceeds the buffer length, EMSGSIZE is
returned. This commit adds code to optionally check for that buffer and chunk
reads and writes across multiple calls to ensure that the buffer is never
exceeded.